--- flink-libraries/flink-table/src/main/scala/org/apache/flink/api/table/plan/nodes/logical/FlinkJoin.scala
@@ -1,46 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one
- * or more contributor license agreements.  See the NOTICE file
- * distributed with this work for additional information
- * regarding copyright ownership.  The ASF licenses this file
- * to you under the Apache License, Version 2.0 (the
- * "License"); you may not use this file except in compliance
- * with the License.  You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.apache.flink.api.table.plan.nodes.logical
-
-import org.apache.calcite.plan.{RelOptCluster, RelTraitSet}
-import org.apache.calcite.rel.RelNode
-import org.apache.calcite.rel.core.{JoinRelType, Join}
-import org.apache.calcite.rex.RexNode
-
-class FlinkJoin(
-    cluster: RelOptCluster,
-    traitSet: RelTraitSet,
-    left: RelNode,
-    right: RelNode,
-    condition: RexNode,
-    joinType: JoinRelType,
-    variablesStopped: java.util.Set[String])
-  extends Join(cluster, traitSet, left, right, condition, joinType, variablesStopped)
-  with FlinkRel {
-
-  override def copy(
-      traitSet: RelTraitSet,
-      condition: RexNode,
-      left: RelNode,
-      right: RelNode,
-      joinType: JoinRelType,
-      semiJoinDone: Boolean): Join = {
-    new FlinkJoin(cluster, traitSet, left, right, condition, joinType, getVariablesStopped)
-  }
-}
